+++ src/main/java/com/busanit501/travelproject/repository/reservation/ReservationByOrderRepositoryImpl.java
@@ -1,20 +1,17 @@
-package com.busanit501.travelproject.repository;
+package com.busanit501.travelproject.repository.reservation;
 
-import com.busanit501.travelproject.domain.QMember;
 import com.busanit501.travelproject.domain.QReservation;
 import com.busanit501.travelproject.domain.Reservation;
 import com.busanit501.travelproject.domain.common.ReservationOrder;
-import com.busanit501.travelproject.dto.ReservationDTO;
+import com.busanit501.travelproject.dto.reservation.ReservationDTO;
 import com.querydsl.core.BooleanBuilder;
 import com.querydsl.core.types.Projections;
 import com.querydsl.jpa.JPQLQuery;
-import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.support.QuerydslRepositorySupport;
-import org.springframework.stereotype.Repository;
 
 import java.util.List;
