+++ src/main/java/com/busanit501/travelproject/repository/reservation/ReservationRepository.java
@@ -1,14 +1,7 @@
-package com.busanit501.travelproject.repository;
+package com.busanit501.travelproject.repository.reservation;
 
 import com.busanit501.travelproject.domain.Reservation;
-import com.busanit501.travelproject.domain.common.ReservationOrder;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
-
-import java.util.List;
 
 public interface ReservationRepository extends JpaRepository<Reservation, Long>, ReservationByOrderRepository {
     // insert delete 있음
