+++ src/main/java/com/busanit501/travelproject/service/reservation/ReservationServiceImpl.java
@@ -1,13 +1,12 @@
-package com.busanit501.travelproject.service;
+package com.busanit501.travelproject.service.reservation;
 
 import com.busanit501.travelproject.domain.Reservation;
 import com.busanit501.travelproject.domain.common.ReservationOrder;
-import com.busanit501.travelproject.dto.ReservationDTO;
-import com.busanit501.travelproject.dto.util.HcbPageRequestDTO;
-import com.busanit501.travelproject.dto.util.HcbPageResponseDTO;
-import com.busanit501.travelproject.dto.util.PageRequestDTO;
-import com.busanit501.travelproject.dto.util.PageResponseDTO;
-import com.busanit501.travelproject.repository.ReservationRepository;
+import com.busanit501.travelproject.dto.reservation.ReservationDTO;
+import com.busanit501.travelproject.dto.util.reservationPageDTO.HcbPageRequestDTO;
+import com.busanit501.travelproject.dto.util.reservationPageDTO.HcbPageResponseDTO;
+import com.busanit501.travelproject.repository.member.MemberRepository;
+import com.busanit501.travelproject.repository.reservation.ReservationRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
 import org.modelmapper.ModelMapper;
@@ -19,11 +18,18 @@ import org.springframework.stereotype.Service;
 public class ReservationServiceImpl implements ReservationService{
     private final ReservationRepository reservationRepository;
     private final ModelMapper modelMapper;
+    private final MemberRepository memberRepository;
 
     @Override
     public Long registerReservation(ReservationDTO reservationDTO) {
-        Reservation reservation = modelMapper.map(reservationDTO, Reservation.class);
-        return 0L;
+        Reservation reservation = Reservation.builder()
+                .reservationNo(reservationDTO.getReservationNo())
+                .member(memberRepository.findByMemberNo(reservationDTO.getMemberNo()))
+                // product 추가해야함
+                .reservationOrder(reservationDTO.getReservationOrder())
+                .build();
+        Reservation result = reservationRepository.save(reservation);
+        return result.getReservationNo();
     }
 
     @Override
